configuration-MassTransit[B22]: State machine configuration

+++ OrdersMS/OrdersMS/src/Contracts/Application/Types/GetPolicyResponse.cs
@@ -4,9 +4,9 @@
     (
         string Id,
         string Type,
-        int CoverageKm,
-        double CoverageAmount,
-        double PriceExtraKm,
+        decimal CoverageKm,
+        decimal CoverageAmount,
+        decimal PriceExtraKm,
         bool IsActive
     );
 }
